Feature and bug: Add bash files
Try to automate the syncing between the Excel file and Calendar through a bash scrip that runs in the background and triggers the execution of main.py every time there is a change in the sheet. However, it isn't possible in Linux (haven't tried in Mac), because it requires the scheduling of the task automatically every time the computer starts and because multiple tasks of the script were suddenly running. However, a run_manually.sh script was added to run the Python file more easily.